+++ change_log.md
@@ -2,6 +2,11 @@
 1.0.x are builds for KSP 1.3.x  
 1.1.x are builds for KSP 1.4.x  
 
+1.0.3 & 1.1.3
+Group Actions; You can now select multiple craft (hold ctrl while selecting) and then perform actions across all of them; add/remove Tags, Transfer between craft type [SPH/VAB/Subs], Move/Copy between saves and Delete.
+Fix: version for KSP 1.4.x was not detecting ctrl + click (previously used for selecting multiple craft type filters, now also used in group selecting craft)
+
+
 1.0.2 & 1.1.2
 Fix: issue caused when creating list of installed parts when there are more than 1 parts with the same name.
 New: CM can use KerbalX to lookup which mods missing parts on a craft belong to.